Refactor worker kickoff
Removes duplicated code across workers.
